08/16
COMPLETED PRO WITH LOGIN AND DATABASE

+++ frontend/src/component/ProductDetails.js
@@ -1,12 +1,19 @@
 import { useProductsContext } from '../hooks/useProductsContext';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 const ProductDetails = ({ product }) => {
   const { dispatch } = useProductsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch('/api/products/' + product._id, {
       method: 'DELETE',
+      headers: { Authorization: `Bearer ${user.token}` },
     });
     const json = await response.json();
